Split nuxt.config.js into managable files in config/nuxt/ dir
