+++ config/nuxt/css.js
@@ -1,0 +1,6 @@
+/**
+ * @see https://nuxtjs.org/api/configuration-css
+ */
+module.exports = [
+  'normalize.css',
+]
