+++ config/nuxt/generate.js
@@ -1,0 +1,9 @@
+const routes = require('./lib/routes')
+
+/**
+ * @see https://nuxtjs.org/api/configuration-generate
+ */
+module.exports = {
+  dir: 'dist/client/',
+  routes,
+}
