+++ config/nuxt/lib/data.js
@@ -1,0 +1,9 @@
+const appConfig = require('../../../src/client/static/data/app.json')
+const locales   = require('../../../src/client/static/data/locales.json')
+const pages     = require('../../../src/client/static/data/pages.json')
+
+module.exports = {
+  appConfig,
+  locales,
+  pages,
+}
