+++ config/nuxt/loading.js
@@ -1,0 +1,6 @@
+/**
+ * @see https://nuxtjs.org/api/configuration-loading
+ */
+module.exports = {
+  color: '#3B8070', // progress bar color
+}
