+++ config/nuxt/readme.md
@@ -1,0 +1,3 @@
+# Nuxt configuration
+
+This directory contains the [Nuxt configuration](https://nuxtjs.org/guide/configuration/). Larger config properties are split into separate `.js` files, named after the config property.
