+++ config/nuxt/router.js
@@ -1,0 +1,6 @@
+/**
+ * @see https://nuxtjs.org/api/configuration-router
+ */
+module.exports = {
+  middleware: ['enforce-trailing-slash'],
+}
