+++ config/readme.md
@@ -1,0 +1,3 @@
+# Configuration
+
+This directory contains configuration files for this development setup.
