+++ config/storybook/readme.md
@@ -1,0 +1,9 @@
+# Storybook configuration
+
+> Storybook is a UI development environment for your UI components. With it, you can visualize different states of your UI components and develop them interactively.
+>
+> Storybook runs outside of your app. So you can develop UI components in isolation without worrying about app specific dependencies and requirements.
+>
+> — [Storybook introduction](https://storybook.js.org/basics/introduction/)
+
+This directory contains the configuration for [Storybook for Vue](https://storybook.js.org/basics/guide-vue/#docs-content).
